requirements.txt
discord.py
requests
python-dotenv
aiosqlite
